+++ cmake/third_party.cmake
@@ -1,0 +1,9 @@
+include(cmake/CPM.cmake)
+
+CPMAddPackage(
+    NAME spdlog
+    GITHUB_REPOSITORY gabime/spdlog
+    GIT_TAG v1.15.0
+)
+
+
