+++ source/parser/ast.cpp
@@ -1,4 +1,4 @@
-#include "ast.hpp"
+#include "parser/ast.hpp"
 
 #include <iostream>
 #include <filesystem>
